requirements.txt
accelerate==0.24.1
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
diskcache==5.6.3
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
grpcio==1.59.3
grpcio-tools==1.59.3
huggingface-hub==0.19.4
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
numpy==1.26.2
packaging==23.2
pandas==2.1.3
pgvector==0.2.3
Pillow==10.1.0
protobuf==4.25.1
psutil==5.9.6
psycopg==3.1.13
psycopg-binary==3.1.13
psycopg-pool==3.2.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
six==1.16.0
sympy==1.12
tokenizers==0.15.0
tqdm==4.66.1
transformers==4.35.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
yarl==1.9.3
--extra-index-url https://download.pytorch.org/whl/cu116
torch==1.13.1+cu116
torchvision==0.14.1+cu116 
torchaudio==0.13.1 